requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.43.3
torch>=2.2.0
accelerate>=0.33.0
prometheus-client==0.20.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
streamlit==1.37.1
requests==2.32.3
pytest==8.3.2
httpx==0.27.2
pandas==2.2.2
